# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MLK Pro API",
    "version": "v1",
    "description": "Auto-generated from routes/api.php (internal API v1)."
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Local dev"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/settings/company": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "CompanySettingsController@edit",
        "operationId": "CompanySettingsController_edit_get",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\Settings\\CompanySettingsController@edit",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete"
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "CompanySettingsController@update",
        "operationId": "CompanySettingsController_update_put",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\Settings\\CompanySettingsController@update",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete"
        ]
      }
    },
...
  }
}